Set up types
