+++ apps/backend/supabase/functions/api/deno.json
@@ -1,0 +1,3 @@
+{
+  "imports": {}
+}
